refactor: 버킷 생성 페이지 폴더 구조 수정 및 코드 분리
